+++ src/test/java/addressbook/AddressBookTest.java
@@ -52,23 +52,23 @@ public class AddressBookTest {
     public void calculatesAgeDifferenceWithinSameYear() {
         AddressBook addressBook = new AddressBook(new AddressBookLoader(new Clock(), "TestAddressBook.txt", getClass()));
 
-        int difference = addressBook.ageDifferenceInDays(person -> person.name().equals("Jaroslaw Pawlak"),
+        long difference = addressBook.ageDifferenceInDays(person -> person.name().equals("Jaroslaw Pawlak"),
                                                          person -> person.name().equals("Maciej Kowalski"));
 
-        assertThat(difference, equalTo(28 + 31));
+        assertThat(difference, equalTo(28L + 31L));
     }
 
     @Test
     public void calculatesAgeDifferenceForPeopleBornInDifferentYears() {
         AddressBook addressBook = new AddressBook(new AddressBookLoader(new Clock(), "TestAddressBook.txt", getClass()));
 
-        int difference = addressBook.ageDifferenceInDays(person -> person.name().equals("Sarah Stone"),
+        long difference = addressBook.ageDifferenceInDays(person -> person.name().equals("Sarah Stone"),
                                                          person -> person.name().equals("Paul Robinson"));
 
         // days between 20/09/80 and 15/01/85
-        int daysIn1980 = 10 + 31 + 30 + 31;
-        int daysIn1981_1984 = 365 + 365 + 365 + 366;
-        int daysIn1985 = 15;
+        long daysIn1980 = 10 + 31 + 30 + 31;
+        long daysIn1981_1984 = 365 + 365 + 365 + 366;
+        long daysIn1985 = 15;
         assertThat(difference, equalTo(daysIn1980 + daysIn1981_1984 + daysIn1985));
     }
